Misc/requirements.txt
numpy
scipy
numpy-stl
scikit-learn
easygui
pymeshlab